refactor(cron): extract queue processing logic to service layer
- Move business logic from cron API route to evaluateQueueService.js
- Add processEvaluationQueue() and processEvaluationTask() methods
- Simplify cron route handler to focus on auth and delegation
